style: update sidebar transition and disable the overflow-x scrollbar
